Allow specifying custom GOPATH
This change adds an '--go-path' option for overriding the default
GOPATH setting. Normally GOPATH is set to some directory in /tmp, so
go get will download the project and its dependencies. This is good
for isolation and reproducibility, but inconvenient. Overriding GOPATH
with --go-path option allows user to reuse already downloaded project
and its dependencies to create ACI. Another upside is creating ACI
from a local version of a project without pushing changes to
repository.