Put asset code in proj2aci package
This rearranges the code a bit to make it possible to read it from top
to bottom.

This also adds a code trying to copy shared libraries required by an
asset being a library or executable. That way we can specify an asset
like '/usr/sbin/iptables:/usr/sbin/iptables' and it will be copied
along with several libraries it needs.

Unfortunately, the libraries loaded at application runtime pose a
problem. An "ldd" (or "objdump") won't show them as needed, so we
don't copy them. For now I only hardcoded glibc's name service switch
(libnss_*.so) libraries which are needed for networking.